Decompose prediction into one function per file
predict keeps the orchestration; per-workflow expansion, the cached ref
resolver, the cached workflow fetcher, and the granted executor factory each
move to their own module under src/predict/.
